+++ rog-control-center/src/system_state.rs
@@ -6,9 +6,10 @@ use std::time::SystemTime;
 use egui::Vec2;
 use log::error;
 use rog_anime::{Animations, DeviceState};
+use rog_aura::aura_detection::PowerZones;
 use rog_aura::layouts::KeyLayout;
-use rog_aura::usb::AuraPowerDev;
-use rog_aura::{AuraEffect, AuraModeNum, LedBrightness};
+use rog_aura::usb::{AuraDevice, AuraPowerDev};
+use rog_aura::{AuraEffect, AuraModeNum, AuraZone, LedBrightness};
 use rog_platform::platform::{GpuMode, PlatformPolicy};
 use rog_profiles::fan_curve_set::CurveData;
 use rog_profiles::FanCurvePU;
@@ -110,6 +111,9 @@ pub struct AuraState {
     pub current_mode: AuraModeNum,
     pub modes: BTreeMap<AuraModeNum, AuraEffect>,
     pub enabled: AuraPowerDev,
+    pub dev_type: AuraDevice,
+    pub supported_basic_zones: Vec<AuraZone>,
+    pub supported_power_zones: Vec<PowerZones>,
     /// Brightness from 0-3
     pub bright: LedBrightness,
     pub wave_red: [u8; 22],
@@ -132,6 +136,17 @@ impl AuraState {
                 BTreeMap::new()
             },
             enabled: dbus.proxies().aura().led_power().unwrap_or_default(),
+            supported_basic_zones: dbus
+                .proxies()
+                .aura()
+                .supported_basic_zones()
+                .unwrap_or_default(),
+            supported_power_zones: dbus
+                .proxies()
+                .aura()
+                .supported_power_zones()
+                .unwrap_or_default(),
+            dev_type: dbus.proxies().aura().device_type().unwrap_or_default(),
             bright: dbus.proxies().aura().brightness().unwrap_or_default(),
             wave_red: [0u8; 22],
             wave_green: [0u8; 22],